some changes

+++ src/main/java/ifrs/edu/Control/ControlDelete.java
@@ -7,7 +7,7 @@ import ifrs.edu.Entity.User;
 
 public class ControlDelete {
 
-    public static Response controlDeleteUser(int id){
+    public static Response controlDeleteUser(Long id){
         User user = findUserById(id);
         if(user == null) return Response.status(404).build();
 
@@ -15,11 +15,11 @@ public class ControlDelete {
         return Response.ok().build();
     }
 
-    public static User findUserById(int id){
+    public static User findUserById(Long id){
         return User.findById(id);
     }
 
-    public static Response controlDeleteTrick(int id) {
+    public static Response controlDeleteTrick(Long id) {
         Trick trick = findTrickById(id);
         if(trick == null) return Response.status(404).build();
 
@@ -27,7 +27,7 @@ public class ControlDelete {
         return Response.ok().build();
     }
 
-    public static Trick findTrickById(int id){
+    public static Trick findTrickById(Long id){
         return Trick.findById(id);
     }
